changed Duke to Goot
